% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luscSubset.R
\docType{data}
\name{luscSubset}
\alias{luscSubset}
\title{DNA copy number data for lung squamous cell carcinoma.}
\format{
A numeric matrix of quantitative DNA copy number values with 3480 rows and 60 columns.
Rows are indexed by genes, and columns are indexed by samples.  The entries are on the log
ratio scale and were produced by the GISTIC pipeline.
}
\source{
\url{https://gdac.broadinstitute.org/}
}
\usage{
luscSubset
}
\description{
A subset of the DNA copy number data from the TCGA lung
lung squamous cell carcinoma cohort.
}
\details{
The original data set downloaded from \url{https://gdac.broadinstitute.org/}
contained gene-level segmented DNA copy number values for 24776 genes and 501 samples
that were produced by the GISTIC pipeline.  Because of the size limitations on data
in R packages, a random subset of genes and samples was selected.  By construction,
the number of genes and samples in the lung squamous cell carcinoma data is different
than the number of genes and samples in the lung adenocarcinoma data.
}
\keyword{datasets}
