% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pD.R
\docType{data}
\name{pD}
\alias{pD}
\title{Prepped data for DiNAMIC.Duo.}
\format{
A list containing three components for a common set of genes.
}
\usage{
pD
}
\description{
A list produced by the dataPrep() function.
}
\details{
The components of the list are named X, Y, and posDT.  They contain the DNA
copy number data for lung adenocarcinoma, the DNA copy number data for lung squamous
cell carcinoma, and gene position data, respectively.  This data object was produced
by applying \code{\link{dataPrep}} using the \code{\link{luadSubset}} and \code{\link{luscSubset}}
data sets.  This reduces run time when the package is compiled by CRAN, thus eliminating run
time errors.
}
\keyword{datasets}
